Ported GPU solver into the OOP gramework

+++ include/VtkWriter.hpp
@@ -5,9 +5,9 @@
 #include <string>
 
 namespace LBM {
-	void WriteVtk(const std::vector<double>& rho, 
-				   const std::vector<double>& ux, 
-				   const std::vector<double>& uy, 
+	void WriteVtk(const std::vector<float>& rho, 
+				   const std::vector<float>& ux, 
+				   const std::vector<float>& uy, 
 				   const size_t Nx, const size_t Ny, 
 				   const std::string& file_path, 
 				   const std::string& pv_title);
